requirements.txt
# Core dependencies
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
Pillow>=9.0.0

# Existing game engine rendering (optional — not needed for headless RL)
pygame>=2.1.0

# Testing
pytest>=7.0.0

# Legacy RL code (optional)
# torch>=1.10.0
# torchvision>=0.11.0
# gym>=0.21.0
